Update service_decoration.rst
Make it clear that the first paragraph is *not* about decorating a service, but completely overriding it. Decorating is explained in the next paragraph, so the reference to the decorator pattern has been moved there. Otherwise it's easy to get confused when reading the doc.

The text was clear at the beginning, but then the meaning has been changed when the file was moved, in commit https://github.com/symfony/symfony-docs/commit/141e922e0c1a1bf3d0b62cc8bf343ccb848f71a9#diff-ebbbc2e2adacde7254ce22cf2a32198c